[FLINK-16200][table-planner] Refactor CustomizedConvertRule
